Add files via upload
I had initially thought about running the FTP portion of the bot as a thread in the code. This was a test run that I wrote in order to better understand how to properly implement such a function. I didn't end up using a threading system, however, because I found that the discord.py library had a function that allowed for background operating functions.